first working example

+++ src/gdextension_api_Glyph.v
@@ -2,16 +2,16 @@ module vgdextension
 
 pub struct Glyph {
     pub mut:
-    start int = -1
-    end int = -1
+    start i32 = -1
+    end i32 = -1
     count u8
     repeat u8 = 1
     flags u16
-    x_off f32
-    y_off f32
-    advance f32
+    x_off f64
+    y_off f64
+    advance f64
     font_rid RID
-    font_size int
+    font_size i32
     index i32
 }
